+++ common/src/main/java/com/faboslav/friendsandfoes/client/render/entity/feature/MoobloomFlowerFeatureRenderer.java
@@ -4,6 +4,10 @@ import com.faboslav.friendsandfoes.entity.MoobloomEntity;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.PlantBlock;
+import net.minecraft.block.TallFlowerBlock;
+import net.minecraft.block.TallPlantBlock;
+import net.minecraft.block.enums.DoubleBlockHalf;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.VertexConsumerProvider;
@@ -39,7 +43,14 @@ public final class MoobloomFlowerFeatureRenderer<T extends MoobloomEntity> exten
 	) {
 		if (moobloomEntity.isBaby() == false && moobloomEntity.isInvisible() == false) {
 			BlockRenderManager blockRenderManager = MinecraftClient.getInstance().getBlockRenderManager();
-			BlockState blockState = moobloomEntity.getVariant().getFlowerBlock().getDefaultState();
+
+			PlantBlock flower = moobloomEntity.getVariant().getFlower();
+			BlockState blockState = moobloomEntity.getVariant().getFlower().getDefaultState();
+
+			if (flower instanceof TallFlowerBlock || flower instanceof TallPlantBlock) {
+				blockState = moobloomEntity.getVariant().getFlower().getDefaultState().with(TallPlantBlock.HALF, DoubleBlockHalf.UPPER);
+			}
+
 			int overlay = LivingEntityRenderer.getOverlay(moobloomEntity, 0.0F);
 
 			MinecraftClient minecraftClient = MinecraftClient.getInstance();